[🚀 사이클2 - 미션 (예약 변경/취소와 에러 처리)] 테리 미션 제출합니다. (#484)
* test: API 테스트의 역직렬화 대상을 도메인에서 DTO로 변경

- API 테스트가 실제 컨트롤러의 응답 스펙과 일치하도록 검증 객체를 DTO로 변경함.
- 불필요한 도메인 의존성을 제거함.

* test: 테스트 클래스의 접근 제어자 제거

* refactor(service): 도메인 → DTO 변환 로직 중복 제거

* refactor: 사용되지 않는 필드 제거

* fix: 시간대 추가 폼 및 목록에 누락된 종료 시간 필드 반영

* refactor: 무의미한 도메인 DTO 제거

* feat: 지나간 날짜/시간에 대한 예약 생성 불가 처리

* feat: 중복 예약 불가 처리

* feat: 예약이 존재하는 시간 삭제 불가 처리

* fix: 지나간 시간에 대한 예약 처리 로직 수정

- 지나간 시간을 계산할 때 날짜까지 함께 고려하도록 수정함.

* feat: 유효하지 않은 입력값 처리

* feat: 에러 응답 처리

- 발생한 에러를 잡아서 ErrorResponse 형식으로 body를 생성해 응답 코드와 함께 반환하는 GlobalExceptionHandler 추가
- 에러 응답에 대해 사용자가 이해할 수 있도록 브라우저 화면에 표시되는 기능 추가

* feat: 이름으로 본인의 예약 목록 조회하는 기능 추가

* refactor: PageController의 RequestMapping 양식 통일

- RequestMapping을 "/page"로 통일함.

* feat: 사용자의 본인 예약 취소 기능 추가

* feat: 사용자가 본인 예약의 날짜와 시간을 변경하는 기능 추가

* feat: 사용자가 이미 지난 예약을 취소하려고 할 때 에러 처리

* feat: 사용자가 이미 지난 예약을 변경하려고 할 때 에러 처리

* feat: 추가된 기능에 따른 프론트 화면 추가

- 내 예약 목록 페이지 추가

* test: API 정상 흐름 통합 테스트 추가

* refactor: 에러 응답 세분화

- 에러 상황의 상태 코드와 클라이언트 메시지 정보를 정의하는 ErrorCode 추가
- 커스텀 예외 추가
- 추가된 예외에 따라 GlobalExceptionHandler 메서드 추가

* refactor: 에러 응답 세분화 내용 프론트 반영

* docs(README): 에러 응답 명세 추가

* refactor: TimeConfig를 통한 Clock 주입 제거

- 테스트 코드 상에서의 날짜 변경을 위해 Clock을 주입 받던 코드를 제거하고, 더미데이터(popular-themes-test-data.sql)에서 DATEADD를 사용해 데이터를 관리하도록 변경.

* test(service): 비즈니스 정책 예외 상황이 포함된 테스트 코드 추가

* fix: Nullable한 필드에 있던 @NotNull 어노테이션 제거

* refactor: 검증 메서드명이 명확한 의미를 담도록 수정

* feat: 잔여 예약이 존재하는 테마 삭제 시도 시 에러 처리

* refactor: 예약 변경 정보를 Request Body로 받아오도록 변경

+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,12 +3,13 @@ package roomescape.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Theme;
-import roomescape.domain.dto.ThemeCreateData;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.ThemeException;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.dto.request.ThemeCreateRequest;
-import roomescape.service.dto.response.ThemeResponse;
+import roomescape.service.dto.command.ThemeCreateCommand;
+import roomescape.service.dto.result.ThemeResult;
 
-import java.time.Clock;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -17,61 +18,58 @@ import java.util.List;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
-    private final Clock clock;
+    private final ReservationRepository reservationRepository;
 
     private static final int DATA_RANGE = 7;
 
-    public ThemeResponse create(final ThemeCreateRequest request) {
+    public ThemeResult create(final ThemeCreateCommand request) {
         final Theme themeWithoutId = Theme.create(
-                new ThemeCreateData(
-                        request.name(),
-                        request.description(),
-                        request.thumbnailUrl()
-                )
+                request.name(),
+                request.description(),
+                request.thumbnailUrl()
         );
 
         Theme theme = themeRepository.save(themeWithoutId);
 
-        return new ThemeResponse(
+        return mapDomainToDto(theme);
+    }
+
+    public void delete(final Long themeId) {
+        final boolean hasAnyOngoingReservation = reservationRepository.existsByThemeId(themeId);
+        if (hasAnyOngoingReservation) {
+            throw new ThemeException(ErrorCode.THEME_HAS_RESERVATION);
+        }
+
+        boolean deleted = themeRepository.deleteById(themeId);
+
+        if (!deleted) {
+            throw new ThemeException(ErrorCode.THEME_NOT_FOUND);
+        }
+    }
+
+    public List<ThemeResult> getPopularThemes() {
+        final LocalDate today = LocalDate.now();
+        final LocalDate startDate = today.minusDays(DATA_RANGE);
+
+        return themeRepository.findPopularThemes(startDate, today)
+                .stream()
+                .map(ThemeService::mapDomainToDto)
+                .toList();
+    }
+
+    public List<ThemeResult> getThemes() {
+        return themeRepository.findAll()
+                .stream()
+                .map(ThemeService::mapDomainToDto)
+                .toList();
+    }
+
+    private static ThemeResult mapDomainToDto(Theme theme) {
+        return new ThemeResult(
                 theme.getId(),
                 theme.getName(),
                 theme.getDescription(),
                 theme.getThumbnailUrl()
         );
-    }
-
-    public void delete(final Long themeId) {
-        boolean deleted = themeRepository.deleteById(themeId);
-
-        if (!deleted) {
-            throw new IllegalArgumentException("존재하지 않는 테마입니다.");
-        }
-    }
-
-    public List<ThemeResponse> getPopularThemes() {
-        final LocalDate today = LocalDate.now(clock);
-        final LocalDate startDate = today.minusDays(DATA_RANGE);
-
-        return themeRepository.findPopularThemes(startDate, today)
-                .stream()
-                .map(theme -> new ThemeResponse(
-                        theme.getId(),
-                        theme.getName(),
-                        theme.getDescription(),
-                        theme.getThumbnailUrl()
-                ))
-                .toList();
-    }
-
-    public List<ThemeResponse> getThemes() {
-        return themeRepository.findAll()
-                .stream()
-                .map(theme -> new ThemeResponse(
-                        theme.getId(),
-                        theme.getName(),
-                        theme.getDescription(),
-                        theme.getThumbnailUrl()
-                ))
-                .toList();
     }
 }
